# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Swim Coach API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BookingService =
  (typeof BookingService)[keyof typeof BookingService];

export const BookingService = {
  private_lesson: "private_lesson",
  group_session: "group_session",
  stroke_clinic: "stroke_clinic",
  video_analysis: "video_analysis",
  package_5: "package_5",
  package_10: "package_10",
} as const;

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Booking {
  id: number;
  name: string;
  email: string;
  phone?: string;
  service: BookingService;
  preferredDate?: string;
  preferredTime?: string;
  notes?: string;
  status: BookingStatus;
  createdAt: string;
}

export type CreateBookingRequestService =
  (typeof CreateBookingRequestService)[keyof typeof CreateBookingRequestService];

export const CreateBookingRequestService = {
  private_lesson: "private_lesson",
  group_session: "group_session",
  stroke_clinic: "stroke_clinic",
  video_analysis: "video_analysis",
  package_5: "package_5",
  package_10: "package_10",
} as const;

export interface CreateBookingRequest {
  name: string;
  email: string;
  phone?: string;
  service: CreateBookingRequestService;
  preferredDate?: string;
  preferredTime?: string;
  notes?: string;
}

export interface ContactRequest {
  name: string;
  email: string;
  message: string;
}

export interface ContactResponse {
  success: boolean;
  message: string;
}

export type ClientStatus = (typeof ClientStatus)[keyof typeof ClientStatus];

export const ClientStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface Client {
  id: number;
  bookingId?: number;
  name: string;
  email: string;
  phone?: string;
  neighborhood?: string;
  service?: string;
  notes?: string;
  status: ClientStatus;
  createdAt: string;
}

export interface CoachingPlan {
  id: number;
  clientId?: number;
  title: string;
  goal?: string;
  drills?: string;
  notes?: string;
  shareToken: string;
  isPublic: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePlanRequest {
  title: string;
  goal?: string;
  drills?: string;
  notes?: string;
  clientId?: number;
  isPublic?: boolean;
}

export interface UpdatePlanRequest {
  title?: string;
  goal?: string;
  drills?: string;
  notes?: string;
  clientId?: number;
  isPublic?: boolean;
}

export interface SharedPlan {
  title: string;
  goal?: string;
  drills?: string;
  notes?: string;
  clientName?: string;
  createdAt: string;
  updatedAt: string;
}

export type SessionStatus = (typeof SessionStatus)[keyof typeof SessionStatus];

export const SessionStatus = {
  scheduled: "scheduled",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Session {
  id: number;
  clientId?: number;
  bookingId?: number;
  scheduledAt?: string;
  durationMinutes?: number;
  service?: string;
  notes?: string;
  status: SessionStatus;
  createdAt: string;
  updatedAt: string;
}

export interface CreateSessionRequest {
  clientId?: number;
  bookingId?: number;
  scheduledAt?: string;
  durationMinutes?: number;
  service?: string;
  notes?: string;
  status?: string;
}

export type InvoiceStatus = (typeof InvoiceStatus)[keyof typeof InvoiceStatus];

export const InvoiceStatus = {
  draft: "draft",
  sent: "sent",
  paid: "paid",
} as const;

export interface Invoice {
  id: number;
  clientId?: number;
  sessionId?: number;
  amountCents: number;
  status: InvoiceStatus;
  dueDate?: string;
  paymentMethod?: string;
  notes?: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateInvoiceRequest {
  clientId?: number;
  sessionId?: number;
  amountCents: number;
  status?: string;
  dueDate?: string;
  paymentMethod?: string;
  notes?: string;
}

export type RejectBookingBody = {
  note?: string;
};

export type UpdateCoachClientBody = {
  notes?: string;
  status?: string;
};

export type GetCoachSessionsParams = {
  clientId?: number;
};

export type UpdateCoachSessionBody = {
  scheduledAt?: string;
  durationMinutes?: number;
  service?: string;
  notes?: string;
  status?: string;
};

export type GetCoachInvoicesParams = {
  clientId?: number;
  status?: string;
};

export type UpdateCoachInvoiceBody = {
  amountCents?: number;
  status?: string;
  dueDate?: string;
  paymentMethod?: string;
  notes?: string;
};