# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Swim Coach API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    get:
      operationId: getBookings
      tags: [bookings]
      summary: Get all bookings (coach-protected)
      parameters:
        - in: header
          name: x-coach-password
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
        "401":
          description: Unauthorized
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
...
components:
  schemas:
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        service:
          type: string
          enum: [private_lesson, group_session, stroke_clinic, video_analysis, package_5, package_10]
        preferredDate:
          type: string
        preferredTime:
          type: string
        notes:
          type: string
        status:
          type: string
          enum: [pending, confirmed, cancelled, approved, rejected]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - service
        - status
        - createdAt
    CreateBookingRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        service:
          type: string
          enum: [private_lesson, group_session, stroke_clinic, video_analysis, package_5, package_10]
        preferredDate:
          type: string
        preferredTime:
          type: string
        notes:
          type: string
      required:
        - name
        - email
        - service
...